[VideoPlayer] CVideoBuffer move to Buffers directory
