+++ xbmc/cores/VideoPlayer/Buffers/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(SOURCES VideoBuffer.cpp)
+set(HEADERS VideoBuffer.h)
+
+core_add_library(videoplayer-buffers)
