+++ xbmc/cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodec.h
@@ -8,10 +8,10 @@
 
 #pragma once
 
-#include "cores/VideoPlayer/Process/ProcessInfo.h"
-#include "cores/VideoPlayer/Process/VideoBuffer.h"
-#include "cores/VideoPlayer/Interface/Addon/DemuxPacket.h"
 #include "DVDResource.h"
+#include "cores/VideoPlayer/Buffers/VideoBuffer.h"
+#include "cores/VideoPlayer/Interface/Addon/DemuxPacket.h"
+#include "cores/VideoPlayer/Process/ProcessInfo.h"
 
 extern "C" {
 #include <libavcodec/avcodec.h>
