+++ xbmc/cores/VideoPlayer/DVDCodecs/Video/VDPAU.h
@@ -26,22 +26,24 @@
  *   of locks needed.
  */
 
+#include "cores/VideoPlayer/Buffers/VideoBuffer.h"
 #include "cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodec.h"
-#include "cores/VideoPlayer/Process/VideoBuffer.h"
-#include <X11/Xlib.h>
-#include <X11/Xutil.h>
-#include "threads/CriticalSection.h"
-#include "threads/SharedSection.h"
 #include "cores/VideoSettings.h"
 #include "guilib/DispResource.h"
+#include "threads/CriticalSection.h"
 #include "threads/Event.h"
+#include "threads/SharedSection.h"
 #include "threads/Thread.h"
 #include "utils/ActorProtocol.h"
 #include "utils/Geometry.h"
+
 #include <deque>
 #include <list>
 #include <map>
 #include <vector>
+
+#include <X11/Xlib.h>
+#include <X11/Xutil.h>
 
 extern "C" {
 #include <libavutil/avutil.h>
