+++ xbmc/cores/VideoPlayer/Process/ProcessInfo.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "VideoBuffer.h"
+#include "cores/VideoPlayer/Buffers/VideoBuffer.h"
 #include "cores/VideoPlayer/VideoRenderers/RenderInfo.h"
 #include "cores/VideoSettings.h"
 #include "threads/CriticalSection.h"
