+++ xbmc/cores/VideoPlayer/Process/gbm/VideoBufferDRMPRIME.h
@@ -8,8 +8,8 @@
 
 #pragma once
 
+#include "cores/VideoPlayer/Buffers/VideoBuffer.h"
 #include "cores/VideoPlayer/DVDCodecs/Video/DVDVideoCodec.h"
-#include "cores/VideoPlayer/Process/VideoBuffer.h"
 
 extern "C"
 {
